Accept removal of "sandboxfs" support in Bazel 7
As of Bazel version 7.0.0, it no longer accepts the
"--experimental_use_sandboxfs" command-line flag, precluding use of
the "sandboxfs" tool to work around kustomize's intolerance for
symbolic links pointing at files outside of the dominating
kustomization root.

Note that this technique is no longer viable, and stop relying on it
in our CI workflow.