Isolate omission of extension to Windows 5.2.1
In version 5.2.1, the released archives for Windows built both for the
AMD64 and ARM64 machine architectures included a single executable
file named "kustomize", with no file extension. In all releases prior
to that one, and in version 5.3.0 released since, that file name is
"kustomize.exe", with the ".exe" file extension common to executable
files for use in Windows.

Ensure that we predict the name of this executable correctly,
introducing a special case for version 5.2.1 that we'll need to
maintain until we drop support for that version.