Use the "rules_shell" Bazle module for "sh_binary"
In order to accommodate the Bazel project's migration away from
loading external repositories automatically for what had been
so-called "native" rules, now defined in their own modules, establish
a dependency on the dedicated external modules and load the files as
necessary to make the definitions available for use.

In our case here, the "sh_binary" rule from the "rules_shell" module
is the only one that requires migration.